+++ .dockerignore
@@ -4,5 +4,6 @@
 credentials.example.json
 node_modules
 Easy\ Setup
+Gruntfile.js
+package.json
 pogom.db
-core
