+++ .gitignore
@@ -8,4 +8,3 @@ config/config.ini
 static/dist/
 node_modules/
 venv/
-core
