--- .landscape.yml
@@ -1,10 +1,0 @@
-
-pylint:
-  disable:
-    # That annoying test which says you should do
-    #   log.info("Step %s", i)
-    # instead of
-    #   log.info("Step {}".format(i))
-    # So you can save a few ms not running string interpolation
-    # at the cost of lacking format() capabilities
-    - W1202
