+++ .travis.yml
@@ -7,9 +7,5 @@ env:
 install:
   - pip install Flask==$FLASK
   - pip install -r requirements.txt
-  - npm install grunt-cli -g
-  - npm install
 script:
-  - npm run lint
-  - npm run build
   - nosetests
