+++ Dockerfile
@@ -17,25 +17,16 @@ ENTRYPOINT ["python", "./runserver.py", "--host", "0.0.0.0"]
 # Set default options when container is run without any command line arguments
 CMD ["-h"]
 
-# Install required system packages
+# add certificates to talk to the internets
 RUN apk add --no-cache ca-certificates
 
+# Copy Python requirements so we only rebuild deps if they have changed
 COPY requirements.txt /usr/src/app/
 
+# Install all prerequisites (build base used for gcc of some python modules)
 RUN apk add --no-cache build-base \
  && pip install --no-cache-dir -r requirements.txt \
  && apk del build-base
 
-COPY package.json Gruntfile.js /usr/src/app/
-COPY static /usr/src/app/static
-
-RUN apk add --no-cache build-base nodejs \
- && npm install -g grunt-cli \
- && npm install \
- && npm run-script build \
- && npm uninstall -g grunt-cli \
- && rm -rf node_modules \
- && apk del build-base nodejs
-
-# Copy everything to the working directory (Python files, templates, config) in one go.
-COPY . /usr/src/app/
+# Add the rest of the app code
+COPY . /usr/src/app
