+++ Gruntfile.js
@@ -4,20 +4,16 @@ module.exports = function(grunt) {
     pkg: grunt.file.readJSON('package.json'),
 
     sass: {
-      dist: {
-        files: {
-          'static/dist/css/app.built.css': [
-            'static/sass/main.scss',
-            'static/sass/pokemon-sprite.scss'
-          ],
-          'static/dist/css/mobile.built.css': 'static/sass/mobile.scss'
-        }
-      }
-    },
+  		dist: {
+  			files: {
+  			  'static/css/main.css' : 'static/sass/main.scss'
+  			}
+  		}
+	  },
     jshint: {
       files: ['Gruntfile.js', 'js/*.js', '!js/vendor/**/*.js'],
       options: {
-        reporter: require('jshint-stylish'),
+		    reporter: require('jshint-stylish'),
         globals: {
           jQuery: true,
           console: true,
@@ -33,54 +29,40 @@ module.exports = function(grunt) {
       },
       dist: {
         files: {
-          'static/dist/js/app.built.js': 'static/js/app.js',
-          'static/dist/js/map.built.js': 'static/js/map.js',
-          'static/dist/js/mobile.built.js': 'static/js/mobile.js',
-          'static/dist/js/stats.built.js': 'static/js/stats.js'
+          'static/dist/js/app.js': 'static/js/app.js',
+          'static/dist/js/map.js': 'static/map.js'
         }
       }
     },
     uglify: {
       options: {
-        banner: '/*\n <%= pkg.name %> <%= grunt.template.today("yyyy-mm-dd") %> \n*/\n',
-        sourceMap: true,
-        compress: {
-          unused: false
-        }
+        banner: '/*\n <%= pkg.name %> <%= grunt.template.today("yyyy-mm-dd") %> \n*/\n'
       },
       build: {
         files: {
-          'static/dist/js/app.min.js': 'static/dist/js/app.built.js',
-          'static/dist/js/map.min.js': 'static/dist/js/map.built.js',
-          'static/dist/js/mobile.min.js': 'static/dist/js/mobile.built.js',
-          'static/dist/js/stats.min.js': 'static/dist/js/stats.built.js'
+          'static/dist/js/app.min.js': 'static/dist/js/app.js'
         }
       }
     },
-    clean: {
-      build: {
-        src: 'static/dist'
-      }
-    },
     watch: {
-      options: {
-        interval: 1000,
-        spawn: true
-      },
-      src: {
-        files: ['**/*.html'],
-        options: { livereload: true }
-      },
-      js: {
-        files: ['**/*.js', '!node_modules/**/*.js', '!static/dist/**/*.js'],
-        options: { livereload: true },
-        tasks: ['js-lint', 'js-build']
-      },
-      css: {
-        files: '**/*.scss',
-        options: { livereload: true },
-        tasks: ['css-build']
-      }
+  		options: {
+  			interval: 1000,
+  			spawn: true
+  		},
+  		src: {
+  			files: ['**/*.html'],
+  			options: { livereload: true }
+  		},
+  		js: {
+  			files: ['**/*.js', '!node_modules/**/*.js', '!static/dist/**/*.js'],
+  			options: { livereload: true },
+        tasks: ['babel', 'uglify']
+  		},
+  		css: {
+  			files: '**/*.scss',
+  			options: { livereload: true },
+  			tasks: ['sass', 'cssmin']
+  		}
     },
     cssmin: {
       options: {
@@ -88,11 +70,13 @@ module.exports = function(grunt) {
       },
       build: {
         files: {
-          'static/dist/css/app.min.css': 'static/dist/css/app.built.css',
-          'static/dist/css/mobile.min.css': 'static/dist/css/mobile.built.css'
+          'static/dist/css/app.min.css': [
+              'static/css/main.css',
+              'static/css/pokemon-sprites.css'
+          ]
         }
       }
-    }
+  	},
   });
 
   grunt.loadNpmTasks('grunt-contrib-uglify');
@@ -107,12 +91,6 @@ module.exports = function(grunt) {
   grunt.loadNpmTasks('grunt-html-validation');
   grunt.loadNpmTasks('grunt-babel');
 
-  grunt.registerTask('js-build', ['babel', 'uglify']);
-  grunt.registerTask('css-build', ['sass', 'cssmin']);
-  grunt.registerTask('js-lint', ['jshint']);
-
-  grunt.registerTask('build', ['clean', 'js-build', 'css-build']);
-  grunt.registerTask('lint', ['js-lint']);
-  grunt.registerTask('default', ['lint', 'build', 'watch']);
+  grunt.registerTask('default', ['jshint', 'sass', 'cssmin', 'babel', 'uglify', 'watch']);
 
 };
