+++ Tools/Hex-Beehive-Generator/beehive_builder.sh
@@ -1,3 +1,8 @@
-#!/usr/bin/env bash
-
-echo "This script isn't required any more, just run 'python location_generator.py'"
+python location_generator.py -st $1 -lp $2 -lat $3 -lon $4 > locations.txt
+echo "" > beehive.sh
+while read line || [[ -n "$line" ]]; do
+	echo "nohup python runserver.py -ns -l '$line' -st $1 &" >> beehive.sh
+done < locations.txt
+echo "nohup python runserver.py -os -l '$3, $4' &" >> beehive.sh
+chmod u+x beehive.sh
+echo "Beehive generated. Run ./beehive.sh to unleash the swarm."
