+++ app.json
@@ -27,13 +27,5 @@
       "description": "Any extra arguments that you want to pass to runserver.py.",
       "required": false
     }
-  },
-  "buildpacks": [
-    {
-      "url": "heroku/nodejs"
-    },
-    {
-      "url": "heroku/python"
-    }
-  ]
+  }
 }
