+++ config/config.ini.example
@@ -11,21 +11,17 @@
 #db-pass:               # required for mysql
 
 # Search settings
-#location:
+#location: 
 #no-gyms:               # disables gym scanning (default false)
 #no-pokemon:            # disables pokemon scanning (default false)
 #no-pokestops:          # disables pokestop scanning (default false)
 #num-threads:           # number of search threads (default 1)
-#scan-delay:            # default 5
+#scan-delay:            # default 1
 #step-limit:            # default 12
 
 # Misc
-#gmaps-key:             # your Google Maps API key
-#webhook:               # webhook URL (including http://)
+#gmaps-key:
 
 # Webserver settings
 #host:                  # address to listen on (default 127.0.0.1)
 #port:                  # port to listen on (default 5000)
-
-#Uncomment a line when you want to change its default value (Remove # at the beginning)
-#username, password, location and gmaps-key are required
