+++ pogom/__init__.py
@@ -5,12 +5,12 @@ config = {
     'LOCALE': 'en',
     'LOCALES_DIR': 'static/locales',
     'ROOT_PATH': '',
-    'DATA_DIR': 'static/data',
     'ORIGINAL_LATITUDE': None,
     'ORIGINAL_LONGITUDE': None,
     'GMAPS_KEY': None,
-    'REQ_SLEEP': 5,
+    'REQ_SLEEP': 1,
     'REQ_HEAVY_SLEEP': 30,
     'REQ_MAX_FAILED': 5,
-    'PASSWORD': None
+    'PASSWORD': None,
+    'SEARCH_QUEUE_DEPTH': 100
 }
