+++ pogom/models.py
@@ -4,98 +4,87 @@
 import logging
 import os
 import time
-from peewee import Model, SqliteDatabase, InsertQuery,\
-                   IntegerField, CharField, DoubleField, BooleanField,\
-                   DateTimeField, OperationalError, create_model_tables
-from playhouse.flask_utils import FlaskDB
-from playhouse.pool import PooledMySQLDatabase
-from playhouse.shortcuts import RetryOperationalError
+from peewee import Model, MySQLDatabase, SqliteDatabase, InsertQuery, IntegerField,\
+                   CharField, DoubleField, BooleanField, DateTimeField,\
+                   OperationalError
 from datetime import datetime, timedelta
 from base64 import b64encode
 
 from . import config
-from .utils import get_pokemon_name, get_pokemon_rarity, get_pokemon_types, get_args, send_to_webhook
+from .utils import get_pokemon_name, get_args, send_to_webhook
 from .transform import transform_from_wgs_to_gcj
 from .customLog import printPokemon
+
+logging.basicConfig(level=logging.INFO, format='%(asctime)s [%(module)11s] [%(levelname)7s] %(message)s')
 
 log = logging.getLogger(__name__)
 
 args = get_args()
-flaskDb = FlaskDB()
+db = None
 
+def init_database():
+    global db
+    if db is not None:
+        return db
 
-class MyRetryDB(RetryOperationalError, PooledMySQLDatabase):
-    pass
-
-
-def init_database(app):
+    print args.db_type
     if args.db_type == 'mysql':
-        db = MyRetryDB(
+        db = MySQLDatabase(
             args.db_name,
             user=args.db_user,
             password=args.db_pass,
-            host=args.db_host,
-            max_connections=args.db_max_connections,
-            stale_timeout=300)
+            host=args.db_host)
         log.info('Connecting to MySQL database on {}.'.format(args.db_host))
     else:
         db = SqliteDatabase(args.db)
         log.info('Connecting to local SQLLite database.')
 
-    app.config['DATABASE'] = db
-    flaskDb.init_app(app)
-
     return db
 
 
-class BaseModel(flaskDb.Model):
+class BaseModel(Model):
+    class Meta:
+        database = init_database()
 
     @classmethod
     def get_all(cls):
         results = [m for m in cls.select().dicts()]
         if args.china:
             for result in results:
-                result['latitude'], result['longitude'] = \
-                    transform_from_wgs_to_gcj(
-                        result['latitude'], result['longitude'])
+                result['latitude'],  result['longitude'] = \
+                    transform_from_wgs_to_gcj(result['latitude'],  result['longitude'])
         return results
-
 
 class Pokemon(BaseModel):
     # We are base64 encoding the ids delivered by the api
     # because they are too big for sqlite to handle
     encounter_id = CharField(primary_key=True, max_length=50)
-    spawnpoint_id = CharField(index=True)
-    pokemon_id = IntegerField(index=True)
+    spawnpoint_id = CharField()
+    pokemon_id = IntegerField()
     latitude = DoubleField()
     longitude = DoubleField()
-    disappear_time = DateTimeField(index=True)
-
-    class Meta:
-        indexes = ((('latitude', 'longitude'), False),)
+    disappear_time = DateTimeField()
 
     @classmethod
     def get_active(cls, swLat, swLng, neLat, neLng):
-        if swLat is None or swLng is None or neLat is None or neLng is None:
+        if swLat == None or swLng == None or neLat == None or neLng == None:
             query = (Pokemon
-                     .select()
-                     .where(Pokemon.disappear_time > datetime.utcnow())
-                     .dicts())
+                 .select()
+                 .where(Pokemon.disappear_time > datetime.utcnow())
+                 .dicts())
         else:
             query = (Pokemon
-                     .select()
-                     .where((Pokemon.disappear_time > datetime.utcnow()) &
-                            (Pokemon.latitude >= swLat) &
-                            (Pokemon.longitude >= swLng) &
-                            (Pokemon.latitude <= neLat) &
-                            (Pokemon.longitude <= neLng))
-                     .dicts())
+                 .select()
+                 .where((Pokemon.disappear_time > datetime.utcnow()) &
+                    (Pokemon.latitude >= swLat) &
+                    (Pokemon.longitude >= swLng) &
+                    (Pokemon.latitude <= neLat) &
+                    (Pokemon.longitude <= neLng))
+                 .dicts())
 
         pokemons = []
         for p in query:
             p['pokemon_name'] = get_pokemon_name(p['pokemon_id'])
-            p['pokemon_rarity'] = get_pokemon_rarity(p['pokemon_id'])
-            p['pokemon_types'] = get_pokemon_types(p['pokemon_id'])
             if args.china:
                 p['latitude'], p['longitude'] = \
                     transform_from_wgs_to_gcj(p['latitude'], p['longitude'])
@@ -105,7 +94,7 @@ class Pokemon(BaseModel):
 
     @classmethod
     def get_active_by_id(cls, ids, swLat, swLng, neLat, neLng):
-        if swLat is None or swLng is None or neLat is None or neLng is None:
+        if swLat == None or swLng == None or neLat == None or neLng == None:
             query = (Pokemon
                      .select()
                      .where((Pokemon.pokemon_id << ids) &
@@ -125,8 +114,6 @@ class Pokemon(BaseModel):
         pokemons = []
         for p in query:
             p['pokemon_name'] = get_pokemon_name(p['pokemon_id'])
-            p['pokemon_rarity'] = get_pokemon_rarity(p['pokemon_id'])
-            p['pokemon_types'] = get_pokemon_types(p['pokemon_id'])
             if args.china:
                 p['latitude'], p['longitude'] = \
                     transform_from_wgs_to_gcj(p['latitude'], p['longitude'])
@@ -140,27 +127,24 @@ class Pokestop(BaseModel):
     enabled = BooleanField()
     latitude = DoubleField()
     longitude = DoubleField()
-    last_modified = DateTimeField(index=True)
-    lure_expiration = DateTimeField(null=True, index=True)
+    last_modified = DateTimeField()
+    lure_expiration = DateTimeField(null=True)
     active_pokemon_id = IntegerField(null=True)
-
-    class Meta:
-        indexes = ((('latitude', 'longitude'), False),)
 
     @classmethod
     def get_stops(cls, swLat, swLng, neLat, neLng):
-        if swLat is None or swLng is None or neLat is None or neLng is None:
+        if swLat == None or swLng == None or neLat == None or neLng == None:
             query = (Pokestop
-                     .select()
-                     .dicts())
+                 .select()
+                 .dicts())
         else:
             query = (Pokestop
-                     .select()
-                     .where((Pokestop.latitude >= swLat) &
-                            (Pokestop.longitude >= swLng) &
-                            (Pokestop.latitude <= neLat) &
-                            (Pokestop.longitude <= neLng))
-                     .dicts())
+                 .select()
+                 .where((Pokestop.latitude >= swLat) &
+                    (Pokestop.longitude >= swLng) &
+                    (Pokestop.latitude <= neLat) &
+                    (Pokestop.longitude <= neLng))
+                 .dicts())
 
         pokestops = []
         for p in query:
@@ -185,25 +169,22 @@ class Gym(BaseModel):
     enabled = BooleanField()
     latitude = DoubleField()
     longitude = DoubleField()
-    last_modified = DateTimeField(index=True)
-
-    class Meta:
-        indexes = ((('latitude', 'longitude'), False),)
+    last_modified = DateTimeField()
 
     @classmethod
     def get_gyms(cls, swLat, swLng, neLat, neLng):
-        if swLat is None or swLng is None or neLat is None or neLng is None:
+        if swLat == None or swLng == None or neLat == None or neLng == None:
             query = (Gym
-                     .select()
-                     .dicts())
+                 .select()
+                 .dicts())
         else:
             query = (Gym
-                     .select()
-                     .where((Gym.latitude >= swLat) &
-                            (Gym.longitude >= swLng) &
-                            (Gym.latitude <= neLat) &
-                            (Gym.longitude <= neLng))
-                     .dicts())
+                 .select()
+                 .where((Gym.latitude >= swLat) &
+                    (Gym.longitude >= swLng) &
+                    (Gym.latitude <= neLat) &
+                    (Gym.longitude <= neLng))
+                 .dicts())
 
         gyms = []
         for g in query:
@@ -211,26 +192,21 @@ class Gym(BaseModel):
 
         return gyms
 
-
 class ScannedLocation(BaseModel):
     scanned_id = CharField(primary_key=True, max_length=50)
     latitude = DoubleField()
     longitude = DoubleField()
-    last_modified = DateTimeField(index=True)
-
-    class Meta:
-        indexes = ((('latitude', 'longitude'), False),)
+    last_modified = DateTimeField()
 
     @classmethod
     def get_recent(cls, swLat, swLng, neLat, neLng):
         query = (ScannedLocation
                  .select()
-                 .where((ScannedLocation.last_modified >=
-                        (datetime.utcnow() - timedelta(minutes=15))) &
-                        (ScannedLocation.latitude >= swLat) &
-                        (ScannedLocation.longitude >= swLng) &
-                        (ScannedLocation.latitude <= neLat) &
-                        (ScannedLocation.longitude <= neLng))
+                 .where((ScannedLocation.last_modified >= (datetime.utcnow() - timedelta(minutes=15))) &
+                    (ScannedLocation.latitude >= swLat) &
+                    (ScannedLocation.longitude >= swLng) &
+                    (ScannedLocation.latitude <= neLat) &
+                    (ScannedLocation.longitude <= neLng))
                  .dicts())
 
         scans = []
@@ -238,7 +214,6 @@ class ScannedLocation(BaseModel):
             scans.append(s)
 
         return scans
-
 
 def parse_map(map_dict, iteration_num, step, step_location):
     pokemons = {}
@@ -253,8 +228,7 @@ def parse_map(map_dict, iteration_num, step, step_location):
                 d_t = datetime.utcfromtimestamp(
                     (p['last_modified_timestamp_ms'] +
                      p['time_till_hidden_ms']) / 1000.0)
-                printPokemon(p['pokemon_data']['pokemon_id'], p['latitude'],
-                             p['longitude'], d_t)
+                printPokemon(p['pokemon_data']['pokemon_id'],p['latitude'],p['longitude'],d_t)
                 pokemons[p['encounter_id']] = {
                     'encounter_id': b64encode(str(p['encounter_id'])),
                     'spawnpoint_id': p['spawnpoint_id'],
@@ -273,40 +247,41 @@ def parse_map(map_dict, iteration_num, step, step_location):
                     'disappear_time': time.mktime(d_t.timetuple())
                 }
 
-                send_to_webhook('pokemon', webhook_data)
+                send_to_webhook('pokemon', webhook_data);
 
-        for f in cell.get('forts', []):
-            if config['parse_pokestops'] and f.get('type') == 1:  # Pokestops
-                    if 'lure_info' in f:
-                        lure_expiration = datetime.utcfromtimestamp(
-                            f['lure_info']['lure_expires_timestamp_ms'] / 1000.0)
-                        active_pokemon_id = f['lure_info']['active_pokemon_id']
-                    else:
-                        lure_expiration, active_pokemon_id = None, None
+        if iteration_num > 0 or step > 50:
+            for f in cell.get('forts', []):
+                if config['parse_pokestops'] and f.get('type') == 1:  # Pokestops
+                        if 'lure_info' in f:
+                            lure_expiration = datetime.utcfromtimestamp(
+                                f['lure_info']['lure_expires_timestamp_ms'] / 1000.0)
+                            active_pokemon_id = f['lure_info']['active_pokemon_id']
+                        else:
+                            lure_expiration, active_pokemon_id = None, None
 
-                    pokestops[f['id']] = {
-                        'pokestop_id': f['id'],
-                        'enabled': f['enabled'],
-                        'latitude': f['latitude'],
-                        'longitude': f['longitude'],
-                        'last_modified': datetime.utcfromtimestamp(
-                            f['last_modified_timestamp_ms'] / 1000.0),
-                        'lure_expiration': lure_expiration,
-                        'active_pokemon_id': active_pokemon_id
-                    }
+                        pokestops[f['id']] = {
+                            'pokestop_id': f['id'],
+                            'enabled': f['enabled'],
+                            'latitude': f['latitude'],
+                            'longitude': f['longitude'],
+                            'last_modified': datetime.utcfromtimestamp(
+                                f['last_modified_timestamp_ms'] / 1000.0),
+                            'lure_expiration': lure_expiration,
+                            'active_pokemon_id': active_pokemon_id
+                        }
 
-            elif config['parse_gyms'] and f.get('type') is None:  # Currently, there are only stops and gyms
-                    gyms[f['id']] = {
-                        'gym_id': f['id'],
-                        'team_id': f.get('owned_by_team', 0),
-                        'guard_pokemon_id': f.get('guard_pokemon_id', 0),
-                        'gym_points': f.get('gym_points', 0),
-                        'enabled': f['enabled'],
-                        'latitude': f['latitude'],
-                        'longitude': f['longitude'],
-                        'last_modified': datetime.utcfromtimestamp(
-                            f['last_modified_timestamp_ms'] / 1000.0),
-                    }
+                elif config['parse_gyms'] and f.get('type') == None:  # Currently, there are only stops and gyms
+                        gyms[f['id']] = {
+                            'gym_id': f['id'],
+                            'team_id': f.get('owned_by_team', 0),
+                            'guard_pokemon_id': f.get('guard_pokemon_id', 0),
+                            'gym_points': f.get('gym_points', 0),
+                            'enabled': f['enabled'],
+                            'latitude': f['latitude'],
+                            'longitude': f['longitude'],
+                            'last_modified': datetime.utcfromtimestamp(
+                                f['last_modified_timestamp_ms'] / 1000.0),
+                        }
 
     pokemons_upserted = 0
     pokestops_upserted = 0
@@ -341,14 +316,10 @@ def parse_map(map_dict, iteration_num, step, step_location):
 
     bulk_upsert(ScannedLocation, scanned)
 
-
-
 def bulk_upsert(cls, data):
     num_rows = len(data.values())
     i = 0
     step = 120
-
-    flaskDb.connect_db()
 
     while i < num_rows:
         log.debug("Inserting items {} to {}".format(i, min(i+step, num_rows)))
@@ -360,15 +331,7 @@ def bulk_upsert(cls, data):
 
         i+=step
 
-    flaskDb.close_db(None)
-
-
 def create_tables(db):
     db.connect()
     db.create_tables([Pokemon, Pokestop, Gym, ScannedLocation], safe=True)
-    db.close()
-
-def drop_tables(db):
-    db.connect()
-    db.drop_tables([Pokemon, Pokestop, Gym, ScannedLocation], safe=True)
     db.close()
