--- scalingo.json
@@ -1,31 +1,0 @@
-{
-  "name": "PokemonGo Map",
-  "description": "Live visualization of all the Pokémon (with option to show gyms and pokéstops) in an area on Pokémon Go.",
-  "repository": "https://github.com/AHAAAAAAA/PokemonGo-Map",
-  "website": "https://jz6.github.io/PoGoMap/",
-  "env": {
-    "AUTH_SERVICE": {
-      "description": "Authentication service to use. (ptc or google)"
-    },
-    "USERNAME": {
-      "description": "Your username to login to the specified authentication service."
-    },
-    "PASSWORD": {
-      "description": "Your password to login to the specified authentication service."
-    },
-    "LOCATION": {
-      "description": "Location on which the map should be centered. This can be an address, co-ordinates or anything that Google Maps accepts."
-    },
-    "GMAPS_KEY": {
-      "description": "A Google Maps API key for this application."
-    },
-    "STEP_COUNT": {
-      "description": "The number of steps to take.",
-      "value": "5"
-    },
-    "EXTRA_ARGS": {
-      "description": "Any extra arguments that you want to pass to runserver.py.",
-      "required": false
-    }
-  }
-}
