--- static/data/mapstyle.json
@@ -1,1 +1,0 @@
-{"roadmap":"Roadmap","satellite":"Satellite","nolabels_style":"No Labels","dark_style":"Dark","style_light2":"Light2","style_pgo":"PokemonGo","dark_style_nl":"Dark (No Labels)","style_light2_nl":"Light2 (No Labels)","style_pgo_nl":"PokemonGo (No Labels)"}
