+++ static/sass/layout/_header.scss
@@ -16,7 +16,6 @@
 		top: 0;
 		width: 100%;
 		z-index: _misc(z-index-base) + 1;
-		@include noSelect();
 
 		> h1 {
 			color: _palette(accent2, fg-bold);
@@ -57,21 +56,6 @@
 				margin-left: 0.5em;
 			}
 
-			&[href="#stats"] {
-				@include icon('\f080');
-				-webkit-tap-highlight-color: rgba(0,0,0,0);
-
-				&:before {
-					margin: 0 0.5em 0 0;
-				}
-			}
-
-			& + a[href="#stats"]:last-child {
-				border-left: solid 1px _palette(accent2, border);
-				padding-left: 1.25em;
-				margin-left: 0.5em;
-			}
-
 			&:last-child {
 				padding-right: 1.25em;
 			}
@@ -80,11 +64,6 @@
 				padding: 0 0.5em;
 
 				& + a[href="#nav"]:last-child {
-					padding-left: 1em;
-					margin-left: 0.25em;
-				}
-				
-				& + a[href="#stats"]:last-child {
 					padding-left: 1em;
 					margin-left: 0.25em;
 				}
