+++ static/sass/libs/_vars.scss
@@ -6,7 +6,6 @@
 // Duration.
 	$duration: (
 		nav:				0.5s,
-		stats:				0.5s,
 		transition:			0.2s
 	);
