+++ static/sass/main.scss
@@ -49,7 +49,6 @@
 
 @import 'layout/header';
 @import 'layout/nav';
-@import 'layout/stats';
 @import 'layout/banner';
 @import 'layout/main';
 @import 'layout/footer';
